Docker container, http cache headers
Default nginx value is 'off', so it doesn't send any headers and the
browser can do whatever it wants.

So send explicit cache header by default, since most of our hosted files
are build with a content hash in the name (specifically to allow good
caching).
But the .html should be not cached, since the 'old' files won't exist
after a new deploy. So trigger a 'must revalidate' for these.