feat(stream-chat): unread messages count, attempt to update the avatar
The unread count mostly works, though sometimes when opening the chat, it still thinks that 1 message is unread.

After user profile change - name or avatar updated - we try to update this data in stream chat but it seems it gets ignored.

World chat is not initialised on start - when widgets are inited