build(plugins): plugin_miro docker build
Allows building a docker image to serve the miro plugin.

Adds craco config for module federation build of plugin_miro.

Change some package.json scripts to use yarn workspace instead of --cwd.
More explicit and with --cwd it can run with the scripts from the root
package.json instead of the ones inside the packages, causing confusing
problems.

For now a multistage docker instead of seperate Dockerfile. Makes it
easier to keep in sync when having multiple plugins.
Add a stage to the build for plugins and refactors the Dockerfile stages
for better caching of these multiple builds.