feat(api): backend api urls with portnumbers
Fix working with backend api urls like "http://localhost:4000".

generatePath function is used on the whole URL. But, like the name
suggests, this is suppoed to only be used on the path of an URL.
Otherwise in case if "http://locahost:3000" :3000 is seen as a path
parameter and will throw an error since no param with the name '3000' is
passed into it.