feat(react-router): override useParams signature to return provided type
const { param } = useParam<{param: string}>();

react-router v6 returns param of type string|undefined - that has to be handled somehow.

We currently decided to put the responsibility of providing correct type to the user of useParams and enforce the type returned by it.

This it can be useParams<{param: string|undefined}>() if needed.

TODO remove param! assertion across the code.