feat(wallets): remember wallet used for sign-in with account, activate it automatically when needed
On different places we can select different wallets and need to be able to use them for transactions, see their balance info, etc.

useStaking plugin can now do it searching for walletConf by address - it has a serious problem though trying to call useWallet of these confs - as they're different, the nested hooks will most certainly break when another address is selected. Right now I use workaround setting the address as a key for parent components but it should be redesigned in future.

Stored wallet info is also used for rendering correct wallet icon for address in the dropdown.

Number of elements are refactored.

It's still not detected when current wallet address differs from required address (selected in dropdown) - it should be.

Selecting wallet if its info is not stored doesn't trigger re-render.