refactor(wallets): rely on wallet internal activation, add callback for the result
We had double activation - in WalletSelectHelper and wallets themselves. It might be causing problems with Talisman wallet activation in prod.

Refactored to use the activation inside useWallet of selected wallet, adding callback reporting success/fail of this operation.

Alternative could be calling activate from WalletLogin, same as in WalletSelectHelper. The problem with it is that some wallets have specifics - like waiting 0.5 seconds before calling activate, perhaps something else. But we could have this 0.5 sec timeout for all of them or implement it inside activate callback function, so it's a valid alternative.