refactor(core3d): implement preview, setting transform, parent as ObjectInstance functions
It allows setting them to both clone and original instances if needed without having it all over the code.

Setting color and texture could also be moved in it in future - right now it's expected that they're only set to the original object, not effect clone.

A potentially better alternative could be having just one instance and set different materials depending on the case - color, texture, transparent or preview effect.

All issues with not-working transform of objects with effects and so on should be fixed now.