feat(object-inspector): separate wrap image and content image, new asset_2d for content
Wrappable image is handled in 3D only and in object-inspector settings.

Content image is handled in Content Viewer (along with text and video) and ObjectInspector Info.

New asset_2d is for Object Content (similar to claimable) and helps differentiate objects with content and the ones with texture at most. For the ones without content we don't show the viewer and don't load the content attributes.